pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "key4ce"
version = "1.0.0"
description = "A minimalist, cross-platform terminal typing game with rich analytics"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Key4ce Team" }
]
keywords = ["typing", "terminal", "game", "tui", "practice"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Education",
]

dependencies = [
    "textual>=0.40.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
key4ce = "key4ce.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["key4ce"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
